pyproject.toml
[project]
name = "repo-factory"
version = "0.1.0"
description = "CLI to generate repositories from YAML configs"
requires-python = ">=3.11"
authors = [{name = "repo-factory", email = "dev@example.com"}]
readme = "README.md"
license = {text = "MIT"}
dependencies = [
  "pyyaml>=6.0",
  "jinja2>=3.1",
  "click>=8.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.8.0",
  "black>=24.10.0",
  "mypy>=1.13",
  "types-PyYAML",
  "types-Jinja2",
  "pre-commit>=3.7"
]

[project.scripts]
repofactory = "repofactory.cli:main"

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120
select = ["E", "F", "B", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/repofactory"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
